Fix crash when there is no queue file

+++ Av1ador/Entry.cs
@@ -283,7 +283,8 @@ namespace Av1ador
         {
             if (System.IO.File.Exists(backup))
                 System.IO.File.Delete(backup);
-            System.IO.File.Move(queue, backup);
+            if (System.IO.File.Exists(queue))
+                System.IO.File.Move(queue, backup);
             var writer = new System.Xml.Serialization.XmlSerializer(typeof(Entry[]));
             var wfile = new StreamWriter(queue);
             Entry[] entries = new Entry[list.Items.Count];
